+++ projects/quang-keycloak/src/lib/quang-keycloak.service.ts
@@ -1,10 +1,11 @@
 import { Injectable, Optional } from '@angular/core'
 import { Store } from '@ngrx/store'
-import { from, Observable, of } from 'rxjs'
+import { from, Observable } from 'rxjs'
 import { KeycloakService } from 'keycloak-angular'
 import { QuangKeycloakConfig } from './quang-keycloak.config'
 import { QuangKeycloakActions } from './quang-keycloak-store/actions'
-
+import { map } from 'rxjs/operators'
+import { QuangKeycloakSelectors } from './quang-keycloak-store/selectors'
 /**
  * service decorator
  */
@@ -89,8 +90,8 @@ export class QuangKeycloakService {
   /**
    * get the user's roles
    */
-  getUserRoles (): Observable<string[]> {
-    return of(this.keyCloak.getUserRoles(true))
+  getUserRoles (): Observable<any> {
+    return from(this.keyCloak.getUserRoles(true))
   }
 
   /**